added two new Env variables for downstream pipeline

+++ kroxylicious-systemtests/src/main/java/io/kroxylicious/systemtests/installation/kroxylicious/CertManager.java
@@ -8,6 +8,10 @@ package io.kroxylicious.systemtests.installation.kroxylicious;
 
 import java.io.IOException;
 
+import io.kroxylicious.systemtests.Environment;
+
+import io.kroxylicious.systemtests.utils.NamespaceUtils;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -42,7 +46,8 @@ public class CertManager {
      */
     public void deploy() {
         LOGGER.info("Deploy cert manager in {} namespace", Constants.CERT_MANAGER_NAMESPACE);
-        if (kubeClient().getNamespace(Constants.CERT_MANAGER_NAMESPACE) != null) {
+        if (kubeClient().getNamespace(Constants.CERT_MANAGER_NAMESPACE) != null
+            || Environment.CERT_MANAGER_INSTALLED.equalsIgnoreCase("true")) {
             LOGGER.warn("Skipping cert manager deployment. It is already deployed!");
             return;
         }
@@ -55,7 +60,12 @@ public class CertManager {
      * @throws IOException the io exception
      */
     public void delete() throws IOException {
+        if (Environment.CERT_MANAGER_INSTALLED.equalsIgnoreCase("true")) {
+            LOGGER.warn("Skipping cert manager deletion. CERT_MANAGER_INSTALLED was set to true");
+            return;
+        }
         LOGGER.info("Deleting Cert Manager in {} namespace", Constants.CERT_MANAGER_NAMESPACE);
         deployment.withGracePeriod(0).delete();
+        NamespaceUtils.deleteNamespaceWithWait(Constants.CERT_MANAGER_NAMESPACE);
     }
 }
